poc/otel-awss3-poc/mock-worker/requirements.txt
# OpenTelemetry SDK
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
